server/requirements.txt
scikit-learn
pandas
numpy
joblib
scipy
mysql-connector-python
python-dotenv
